+++ golem/gui/api.py
@@ -25,7 +25,8 @@ def auth_required(func):
             token = request.headers.get('token', None)
             if token:
                 try:
-                    Users.verify_auth_token(current_app.secret_key, token)
+                    user = Users.verify_auth_token(current_app.secret_key, token)
+                    request.api_user = user
                 except BadSignature:
                     abort(401, 'Token did not match')
                 except SignatureExpired:
@@ -67,12 +68,19 @@ def golem_default_browser():
     return jsonify(session.settings['default_browser'])
 
 
+@api_bp.route('/golem/project-permissions')
+@auth_required
+def golem_permissions_project():
+    return jsonify(Permissions.project_permissions)
+
+
 @api_bp.route('/page/code/save', methods=['PUT'])
 @auth_required
 def page_code_save():
     project = request.json['project']
     page_object_name = request.json['pageName']
     content = request.json['content']
+    _verify_permissions(Permissions.STANDARD, project)
     path = page_object.page_file_path(project, page_object_name)
     page_object.save_page_object_code(project, page_object_name, content)
     _, error = utils.import_module(path)
@@ -84,6 +92,7 @@ def page_code_save():
 def page_delete():
     project = request.json['project']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.ADMIN, project)
     errors = utils.delete_element(project, 'page', full_path)
     return jsonify(errors)
 
@@ -94,6 +103,7 @@ def page_duplicate():
     project = request.json['project']
     full_path = request.json['fullPath']
     new_file_full_path = request.json['newFileFullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     errors = utils.duplicate_element(project, 'page', full_path, new_file_full_path)
     return jsonify(errors)
 
@@ -103,6 +113,7 @@ def page_duplicate():
 def page_elements():
     project = request.args['project']
     page = request.args['page']
+    _verify_permissions(Permissions.READ_ONLY, project)
     result = {
         'error': '',
         'contents': []
@@ -120,6 +131,7 @@ def page_rename():
     project = request.json['project']
     full_filename = request.json['fullFilename']
     new_full_filename = request.json['newFullFilename']
+    _verify_permissions(Permissions.STANDARD, project)
     error = _rename_element(project, full_filename, new_full_filename, 'page')
     return jsonify({'error': error})
 
@@ -132,15 +144,10 @@ def page_save():
     elements = request.json['elements']
     functions = request.json['functions']
     import_lines = request.json['importLines']
+    _verify_permissions(Permissions.STANDARD, project)
     page_object.save_page_object(project, page_object_name, elements, functions,
                                  import_lines)
     return jsonify('page-saved')
-
-
-@api_bp.route('/permissions/project-permissions')
-@auth_required
-def permissions_project():
-    return jsonify(Permissions.project_permissions)
 
 
 @api_bp.route('/project', methods=['POST'])
@@ -148,6 +155,7 @@ def permissions_project():
 def project_create():
     project_name = request.json['project']
     project_name = project_name.strip().replace(' ', '_')
+    _verify_permissions(Permissions.SUPER_USER)
     errors = []
     if len(project_name) < 3:
         errors.append('Project name is too short')
@@ -164,6 +172,7 @@ def project_create():
 @auth_required
 def project_environments():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(environment_manager.get_envs(project))
 
 
@@ -172,6 +181,7 @@ def project_environments():
 def project_environments_save():
     project = request.json['project']
     env_data = request.json['environmentData']
+    _verify_permissions(Permissions.ADMIN, project)
     error = environment_manager.save_environments(project, env_data)
     return jsonify({'error': error})
 
@@ -200,6 +210,7 @@ def project_has_tests():
 @auth_required
 def project_health():
     project = request.args['project']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
     project_data = report_parser.get_last_executions(projects=[project], suite=None,
                                                      limit=1)
     health_data = {}
@@ -221,6 +232,7 @@ def project_page_create():
     project = request.json['project']
     is_dir = request.json['isDir']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     element, errors = _create_project_element(project, 'page', full_path, is_dir)
     return jsonify({'errors': errors, 'element': element})
 
@@ -230,6 +242,7 @@ def project_page_create():
 def project_page_exists():
     project = request.args['project']
     page_name = request.args['page']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(page_object.page_exists(project, page_name))
 
 
@@ -237,6 +250,7 @@ def project_page_exists():
 @auth_required
 def project_page_tree():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(utils.get_pages(project))
 
 
@@ -244,6 +258,7 @@ def project_page_tree():
 @auth_required
 def project_pages():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     path = page_object.pages_base_dir(project)
     page_objects = file_manager.get_files_dot_path(path, extension='.py')
     return jsonify(page_objects)
@@ -255,6 +270,7 @@ def project_suite_create():
     project = request.json['project']
     is_dir = request.json['isDir']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     element, errors = _create_project_element(project, 'suite', full_path, is_dir)
     return jsonify({'errors': errors, 'element': element})
 
@@ -263,6 +279,7 @@ def project_suite_create():
 @auth_required
 def project_suite_tree():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(utils.get_suites(project))
 
 
@@ -270,6 +287,7 @@ def project_suite_tree():
 @auth_required
 def project_supported_browsers():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     settings = settings_manager.get_project_settings(project)
     remote_browsers = settings_manager.get_remote_browser_list(settings)
     default_browsers = gui_utils.get_supported_browsers_suggestions()
@@ -280,6 +298,7 @@ def project_supported_browsers():
 @auth_required
 def project_tags():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(tags_manager.get_project_unique_tags(project))
 
 
@@ -289,6 +308,7 @@ def project_test_create():
     project = request.json['project']
     is_dir = request.json['isDir']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     element, errors = _create_project_element(project, 'test', full_path, is_dir)
     return jsonify({'errors': errors, 'element': element})
 
@@ -297,6 +317,7 @@ def project_test_create():
 @auth_required
 def project_test_tags():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(tags_manager.get_all_project_tests_tags(project))
 
 
@@ -304,6 +325,7 @@ def project_test_tags():
 @auth_required
 def project_test_tree():
     project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
     return jsonify(utils.get_test_cases(project))
 
 
@@ -319,6 +341,7 @@ def report_suite_execution():
     project = request.args['project']
     suite = request.args['suite']
     execution = request.args['execution']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
     execution_data = report_parser.get_execution_data(project=project, suite=suite,
                                                       execution=execution)
     response = jsonify(execution_data)
@@ -328,13 +351,31 @@ def report_suite_execution():
     return response
 
 
-@api_bp.route('/report/suite/last-executions', methods=['POST'])
+@api_bp.route('/report/last-executions')
+@auth_required
+def report_last_executions():
+    user = _get_user_api_or_session()
+    project_list = user.project_list
+    project_data = report_parser.get_last_executions(project_list, limit=5)
+    return jsonify(projects=project_data)
+
+
+@api_bp.route('/report/project/last-executions')
+@auth_required
+def report_project_last_executions():
+    project = request.args['project']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
+    project_data = report_parser.get_last_executions([project], limit=10)
+    return jsonify(projects=project_data)
+
+
+@api_bp.route('/report/suite/last-executions')
 @auth_required
 def report_suite_last_executions():
-    projects = request.json['projects']
-    suite = request.json['suite']
-    limit = request.json['limit']
-    project_data = report_parser.get_last_executions(projects, suite, limit)
+    project = request.args['project']
+    suite = request.args['suite']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
+    project_data = report_parser.get_last_executions([project], suite=suite, limit=50)
     return jsonify(projects=project_data)
 
 
@@ -346,6 +387,7 @@ def report_test_set():
     execution = request.args['execution']
     test_full_name = request.args['testName']
     test_set = request.args['testSet']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
     test_detail = report_parser.get_test_case_data(project, test_full_name,
                                                    suite=suite, execution=execution,
                                                    test_set=test_set, is_single=False,
@@ -363,6 +405,7 @@ def report_test_status():
     project = request.args['project']
     test_case_name = request.args['test']
     timestamp = request.args['timestamp']
+    _verify_permissions(Permissions.REPORTS_ONLY, project)
     path = os.path.join(session.testdir, 'projects', project, 'reports',
                         'single_tests', test_case_name, timestamp)
     result = {
@@ -394,20 +437,39 @@ def report_test_status():
     return jsonify(result)
 
 
-# TODO: split into save_project_settings, save_global_settings
-@api_bp.route('/settings/save', methods=['PUT'])
+@api_bp.route('/settings/global/save', methods=['PUT'])
 @auth_required
-def settings_save():
-    project = request.json['project']
-    project_settings = request.json['projectSettings']
-    global_settings = request.json['globalSettings']
-    settings_manager.save_global_settings(global_settings)
+def settings_global_save():
+    settings = request.json['settings']
+    _verify_permissions(Permissions.SUPER_USER)
+    settings_manager.save_global_settings(settings)
     session.settings = settings_manager.get_global_settings()
-    if project_settings:
-        settings_manager.save_project_settings(project, project_settings)
-        # re-read project settings
-        session.settings = settings_manager.get_project_settings(project)
     return jsonify('settings-saved')
+
+
+@api_bp.route('/settings/global')
+@auth_required
+def settings_global_get():
+    _verify_permissions(Permissions.SUPER_USER)
+    return jsonify(settings_manager.get_global_settings())
+
+
+@api_bp.route('/settings/project/save', methods=['PUT'])
+@auth_required
+def settings_project_save():
+    project = request.json['project']
+    settings = request.json['settings']
+    _verify_permissions(Permissions.ADMIN, project)
+    settings_manager.save_project_settings(project, settings)
+    return jsonify('settings-saved')
+
+
+@api_bp.route('/settings/project')
+@auth_required
+def settings_project_get():
+    project = request.args['project']
+    _verify_permissions(Permissions.READ_ONLY, project)
+    return jsonify(settings_manager.get_project_settings_only(project))
 
 
 @api_bp.route('/suite/delete', methods=['DELETE'])
@@ -415,6 +477,7 @@ def settings_save():
 def suite_delete():
     project = request.json['project']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.ADMIN, project)
     errors = utils.delete_element(project, 'suite', full_path)
     return jsonify(errors)
 
@@ -425,6 +488,7 @@ def suite_duplicate():
     project = request.json['project']
     full_path = request.json['fullPath']
     new_file_full_path = request.json['newFileFullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     errors = utils.duplicate_element(project, 'suite', full_path, new_file_full_path)
     return jsonify(errors)
 
@@ -435,6 +499,7 @@ def suite_rename():
     project = request.json['project']
     full_filename = request.json['fullFilename']
     new_full_filename = request.json['newFullFilename']
+    _verify_permissions(Permissions.STANDARD, project)
     error = _rename_element(project, full_filename, new_full_filename, 'suite')
     return jsonify({'error': error})
 
@@ -444,6 +509,7 @@ def suite_rename():
 def suite_run():
     project = request.json['project']
     suite_name = request.json['suite']
+    _verify_permissions(Permissions.STANDARD, project)
     timestamp = gui_utils.run_suite(project, suite_name)
     return jsonify(timestamp)
 
@@ -458,6 +524,7 @@ def suite_save():
     tags = request.json['tags']
     browsers = request.json['browsers']
     environments = request.json['environments']
+    _verify_permissions(Permissions.STANDARD, project)
     suite_module.save_suite(project, suite_name, test_cases, processes, browsers,
                             environments, tags)
     return jsonify('suite-saved')
@@ -470,6 +537,7 @@ def test_code_save():
     test_case_name = request.json['testName']
     table_test_data = request.json['testData']
     content = request.json['content']
+    _verify_permissions(Permissions.STANDARD, project)
     test_case.save_test_case_code(project, test_case_name, content, table_test_data)
     path = test_case.test_file_path(project, test_case_name)
     _, error = utils.import_module(path)
@@ -481,6 +549,7 @@ def test_code_save():
 def test_delete():
     project = request.json['project']
     full_path = request.json['fullPath']
+    _verify_permissions(Permissions.ADMIN, project)
     errors = utils.delete_element(project, 'test', full_path)
     return jsonify(errors)
 
@@ -491,6 +560,7 @@ def test_duplicate():
     project = request.json['project']
     full_path = request.json['fullPath']
     new_file_full_path = request.json['newFileFullPath']
+    _verify_permissions(Permissions.STANDARD, project)
     errors = utils.duplicate_element(project, 'test', full_path, new_file_full_path)
     return jsonify(errors)
 
@@ -501,6 +571,7 @@ def test_rename():
     project = request.json['project']
     full_filename = request.json['fullFilename']
     new_full_filename = request.json['newFullFilename']
+    _verify_permissions(Permissions.STANDARD, project)
     error = _rename_element(project, full_filename, new_full_filename, 'test')
     return jsonify({'error': error})
 
@@ -513,6 +584,7 @@ def test_run():
     browsers = request.json['browsers']
     environments = request.json['environments']
     processes = request.json['processes']
+    _verify_permissions(Permissions.STANDARD, project)
     timestamp = gui_utils.run_test_case(project, test_name, browsers, environments,
                                         processes)
     return jsonify(timestamp)
@@ -528,6 +600,7 @@ def test_save():
     test_data_content = request.json['testData']
     test_steps = request.json['steps']
     tags = request.json['tags']
+    _verify_permissions(Permissions.STANDARD, project)
     test_case.save_test_case(project, test_name, description, page_objects,
                              test_steps, test_data_content, tags)
     return jsonify('test-saved')
@@ -536,6 +609,7 @@ def test_save():
 @api_bp.route('/users')
 @auth_required
 def users_get():
+    _verify_permissions(Permissions.SUPER_USER)
     users = deepcopy(Users.users())
     for user in users:
         del user['password']
@@ -545,6 +619,7 @@ def users_get():
 @api_bp.route('/users/user')
 @auth_required
 def user_get():
+    _verify_permissions(Permissions.SUPER_USER)
     username = request.args['username']
     user = Users.get_user_dictionary(username)
     if user:
@@ -559,6 +634,7 @@ def users_new():
     email = request.json['email']
     password = request.json['password']
     is_superuser = request.json['isSuperuser']
+    _verify_permissions(Permissions.SUPER_USER)
     project_permissions_raw = request.json['projectPermissions']
     project_permissions = {}
     for project_permission in project_permissions_raw:
@@ -575,6 +651,7 @@ def users_edit():
     email = request.json['email']
     is_superuser = request.json['isSuperuser']
     project_permissions_raw = request.json['projectPermissions']
+    _verify_permissions(Permissions.SUPER_USER)
     project_permissions = {}
     if project_permissions_raw is not None:
         for p in project_permissions_raw:
@@ -586,6 +663,7 @@ def users_edit():
 @api_bp.route('/users/delete', methods=['DELETE'])
 @auth_required
 def users_delete():
+    _verify_permissions(Permissions.SUPER_USER)
     username = request.json['username']
     errors = Users.delete_user(username)
     return jsonify({'errors': errors})
@@ -594,6 +672,7 @@ def users_delete():
 @api_bp.route('/users/reset-password', methods=['POST'])
 @auth_required
 def users_reset_user_pasword():
+    _verify_permissions(Permissions.SUPER_USER)
     username = request.json['username']
     new_password = request.json['newPassword']
     errors = Users.reset_user_password(username, new_password)
@@ -705,3 +784,35 @@ def _create_project_element(project, element_type, full_path, is_dir):
         'is_directory': is_dir
     }
     return element, errors
+
+
+def _get_user_api_or_session():
+    """Get current_user if user is authenticated (Flask Login)
+    or request.api_user if present
+    """
+    user = None
+    if current_user and current_user.is_authenticated:
+        user = current_user
+    elif request.api_user:
+        user = request.api_user
+    return user
+
+
+def _verify_permissions(permission, project=None):
+    """Verify session user or api user has the required permissions.
+    For permission=superuser, project is optional.
+    When permission weight is not reached it will raise HTTP 401.
+    """
+    user = _get_user_api_or_session()
+    if user is None:
+        abort(401)
+    required_permission_weight = Permissions.get_weight(permission)
+    user_weight = 0
+    if user.is_superuser:
+        user_weight = Permissions.get_weight(Permissions.SUPER_USER)
+    elif project:
+        user_weight = user.project_weight(project)
+    else:
+        abort(500, 'project value is required')
+    if required_permission_weight > user_weight:
+        abort(401)